機材情報の別コンポーネント化、機番クリック時に選択機材と機材情報表示
