+++ src/components/pages/ship.tsx
@@ -3,8 +3,7 @@ import React, { useState, useEffect } from 'react';
 import { ShipData } from 'interfaces/index';
 import { ship } from 'lib/api/ship';
 
-import { Box, Card, CardActionArea, CardContent, Typography, Tab} from "@mui/material";
-import { TabContext, TabPanel, TabList} from '@mui/lab'
+import { Box, Card, CardActionArea, CardContent, Typography} from "@mui/material";
 
 import { ShipList } from 'components/utils/ShipList'
 import { ShipDetails } from 'components/pages/ShipDetails'
@@ -12,16 +11,12 @@ import { ShipDetails } from 'components/pages/ShipDetails'
 
 export const Ship: React.FC = () => {
   const [ships, setShips] = useState<ShipData[]>([])
-  const [selectShip, setSelectShip] = useState< string | undefined >("機番を選んで下さい")
-  const [value, setValue] = useState("1")
-
-  const handleChange = (e: React.SyntheticEvent, newValue:string) =>{
-    setValue(newValue)
-  };
+  const [selectShip, setSelectShip] = useState< string | undefined >("")
+  const [isActive, setIsActive] = useState(false)
 
   const handleOnClick = (e: React.MouseEvent<HTMLElement>) => {
-    console.log(e.currentTarget.dataset.ship)
     setSelectShip(e.currentTarget.dataset.ship)
+    setIsActive(true)
 
   };
 
@@ -70,28 +65,7 @@ export const Ship: React.FC = () => {
           )}
         </Box>
       </Box>
-      <Typography sx={{mb:1}}>
-        機材情報
-      </Typography>
-        <Box>
-          <Box sx={{px:2}}>
-            機番 : {selectShip}
-          </Box>
-          <Box sx={{width: '100%', typography:'body1'}}>
-            <TabContext value={value}>
-              <Box sx={{borderBottom:1, borderColor:'divider'}}>
-                <TabList onChange={handleChange} aria-label="tab API Test">
-                  <Tab label="スケジュール" value="1" />
-                  <Tab label="整備" value="2" />
-                  <Tab label="旅客" value="3" />
-                </TabList>
-              </Box>
-              <TabPanel value="1">スケジュール</TabPanel>
-              <TabPanel value="2">整備情報</TabPanel>
-              <TabPanel value="3">旅客情報</TabPanel>
-            </TabContext>
-          </Box>
-        </Box>
+      {isActive? <ShipDetails selectShip={selectShip} /> : "機番を選んで下さい"}
     </>
   )
 }
